combine utils into AuditLogCard.js
