+++ src/Pages/AuditLog/Components/AuditLogCard.js
@@ -1,10 +1,71 @@
-import { getActionDescription } from './getActionDescription';
-import { formatDetails } from './formatDetails';
-import { formatTimestamp } from './formatTimestamp';
-
 const AuditLogCard = ({ log, index }) => {
   const hasDetails = log => {
     return log.details && Object.keys(log.details).length > 0;
+  };
+
+  const formatDetails = details => {
+    if (!details || Object.keys(details).length === 0) {
+      return null;
+    } else {
+      return (
+        <div className='mt-3 p-4 bg-gray-700 rounded-lg border border-gray-600'>
+          <h4 className='font-semibold text-gray-300 mb-2'>Details:</h4>
+          <div className='space-y-1'>
+            {Object.entries(details).map(([key, value]) => (
+              <div key={key} className='text-sm text-gray-300'>
+                <span className='font-medium text-gray-200'>{key}:</span> {String(value)}
+              </div>
+            ))}
+          </div>
+        </div>
+      );
+    }
+  };
+
+  const formatTimestamp = timestamp => {
+    return new Date(timestamp).toLocaleString('en-US', {
+      year: 'numeric',
+      month: 'long',
+      day: 'numeric',
+      hour: '2-digit',
+      minute: '2-digit',
+      second: '2-digit',
+      timeZoneName: 'short',
+    });
+  };
+
+  const simpleActionDescriptions = {
+    SIGN_UP: 'signed up for an account',
+    LOG_IN: 'logged into the system',
+    PRINT_PAGE: 'printed a page',
+    ACCESS_DOOR: 'accessed a door',
+    CREATE_MESSAGE: 'created a message',
+    DELETE_MESSAGE: 'deleted a message',
+  };
+
+  const getActionDescription = log => {
+    const action = log.action;
+
+    // checks if a user updates or deletes ANOTHER user
+    if (action === 'UPDATE_USER') {
+      if (log.documentId && log.documentId !== log.userId) {
+        return "updated another user's account information";
+      }
+      return 'updated their account information';
+    }
+
+    if (action === 'DELETE_USER') {
+      if (log.documentId && log.documentId !== log.userId) {
+        return 'deleted another user account';
+      }
+      return 'deleted their account';
+    }
+
+    if (simpleActionDescriptions[action]) {
+      return simpleActionDescriptions[action];
+    }
+
+    return `performed action: ${action.toLowerCase().replace(/_/g, ' ')}`;
   };
 
   return (
