--- src/Pages/AuditLog/Components/formatTimestamp.js
@@ -1,11 +1,0 @@
-export const formatTimestamp = timestamp => {
-  return new Date(timestamp).toLocaleString('en-US', {
-    year: 'numeric',
-    month: 'long',
-    day: 'numeric',
-    hour: '2-digit',
-    minute: '2-digit',
-    second: '2-digit',
-    timeZoneName: 'short',
-  });
-};
